netdata/packaging/installer: Do not hard fail when a single package is not available
We have seen that in fedora 30, pymongo package got deprecated resulting into the scriptlet failing to
install any of the dependencies, because it aborted the transaction.

To mitigate this, we enforce the --setopt=strict=0 option to allow dnf to continue when a single package fails.
As per the installer workflow, we seem to attempt a best effort installation when something goes wrong, so
this change seems to comply with this. The user will be informed of the failures on the logs and the missing packages
will be identified either during configure if its a hard dependency or during runtime if its a soft one