netdata/packaging: Fix pacman bugs for archlinux
It seems that when testing with docker image archlinux/base, the following bugs were raised
1) For gcc, archlinux comes with a standard gcc runtime package that causes our workflow to falsely think gcc is present.
Added an extra hack for gcc case, to detect specific package presence to assume gcc is present.
Hacky, but i could not find a better/cleaner approach to attend this
2) After some internet digging, found that pacman has a nice --noconfirm flag to skip confirmation.
If we use the previous approach, passing yes to pacman it will break on packages that prompting the user to select package version.
So, to properly mitigate non-interactive mode we use the noconfirm and fallback to the old way for any peculiar situation noconfirm is not present and something fails