Update python package selection for CentOS 8.
CentOS 8 switched from a single `python` package to a `python2` and
`python36` package, with no backwards compatibility package. This
updates the name we use for Python on CentOS 8 to work correctly.

Note that there are currently still other issues with CentOS 8 right now
(namely missing development packages) that preclude Netdata from easily
building or installing on it.